pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "threatlens-ai"
version = "0.1.0"
description = "End-to-end threat intelligence platform with ML-based exploit prediction and RAG-powered remediation"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "shap>=0.45",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "langchain>=0.2",
    "langchain-groq>=0.1",
    "faiss-cpu>=1.8",
    "sentence-transformers>=3.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "joblib>=1.4",
    "python-dotenv>=1.0",
    "pydantic>=2.7",
    "pyarrow>=16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
